% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_plot.R
\name{search_plot}
\alias{search_plot}
\title{search_plot}
\usage{
search_plot(year, search_term, publication = NULL)
}
\arguments{
\item{year}{An search year range with numeric format.}

\item{search_term}{A character string specifying the search term for the PubMed query.}

\item{publication}{A character string specifying the publication name for the PubMed query.abbreviation of the journal name.}
}
\description{
This function searches PubMed for articles matching a given search term. It prints a summary of the search results and the details of the top articles. It also fetches and writes RIS data of the articles to a file.
}
\examples{
search_plot(2015:2019, "phenology", "Nature")
}
